backend/requirements.txt
django==4.2.7
psycopg2-binary==2.9.9
djoser==2.1.0
djangorestframework==3.14.0
django-extra-fields==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
coverage==7.4.1
